pyproject.toml
[project]
name = "folio-mcp"
version = "0.2.1"
description = "MCP server for FOLIO, the Federated Open Legal Information Ontology"
authors = [{ name = "ALEA Institute", email = "hello@aleainstitute.ai" }]
requires-python = ">=3.10,<4.0.0"
readme = "README.md"
license = "MIT"
keywords = ["legal", "ontology", "mcp", "model-context-protocol", "folio", "law"]
dependencies = [
    "mcp[cli]>=1.26",
    "httpx>=0.27",
]

[project.optional-dependencies]
local = ["folio-python[search]>=0.3.3"]

[project.scripts]
folio-mcp = "folio_mcp.server:main"

[project.urls]
Homepage = "https://folio.openlegalstandard.org"
Repository = "https://github.com/alea-institute/folio-mcp"

[dependency-groups]
dev = [
    "pytest>=8.3.1,<9",
    "pytest-asyncio>=0.23.8,<0.24",
    "pytest-cov>=5.0.0,<6",
    "pylint>=3.2.7,<4",
    "black>=24.4.2,<25",
]

[tool.hatch.build.targets.sdist]
include = ["folio_mcp"]

[tool.hatch.build.targets.wheel]
include = ["folio_mcp"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pylint]
max-line-length = 120
disable = [
    "line-too-long",
    "too-few-public-methods",
    "no-self-argument",
]
ignore-patterns = ["test_.*?.py", "conftest.py"]

[tool.isort]
profile = "black"
line_length = 120

[tool.pytest.ini_options]
addopts = "--cov=folio_mcp --cov-report=term-missing"
